WIP: Read document catalog from selfClient
